Fixes a deletion issue when doc_id does not exists
If you try to delete a document with an _id that does not exist, $ES->delete throws an Exception. B/c of default behaviour of reformo/rslim, it bypasses all Exceptions and it returns Exception message. In this case it returns this message as a plain text json that elasticsearch-php returns.